+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project follows
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html) for public releases.
 
+## [1.0.4] - 2026-08-19
+
+### Changed
+
+- Upgrade dependencies
+
 ## [1.0.3] - 2026-07-26
 
 ### Fixed